ai-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
librosa==0.10.1
numpy==1.24.3
scipy==1.11.4
soundfile==0.12.1
google-cloud-speech==2.21.0
opencv-python==4.8.1.78
mediapipe==0.10.7
pydantic==2.5.0
python-dotenv==1.0.0

